Make HBAR a double

+++ project_1/vmc_hybrid/vmc/src/config/constants.h
@@ -1,3 +1,3 @@
 #pragma once
 
-#define HBAR 1
+#define HBAR ((double) 1)
